Add more flexible attributes parameters.
I believe this will allow the attributes to be set easier. A user can simply call the class with the first 4 parameters for all default settings.

If the user chooses to, they can include an array as the 5th parameter to change the attributes given to the database connection.
If the user wants to keep the default attributes but change the charset, they can pass an empty array to the 5th parameter and set the chartet with the 6th parameter.

A user can can also pass NULL (or anything that is not an array) to the 5th parameter to send no attributes to the connection. I wanted this to be an option so people can turn off all attributes instead of it always defaulting to having what I had chosen to be important.

Originally I was going to make it so passing "NULL" to the 5th parameter was the only way to turn off all attributes, but I realized that an empty array also returns as a null, so that was a no go. Do you have any better ideas?